+++ viewer/src/locales/Portuguese.js
@@ -15,219 +15,219 @@
 export default {
     // language info
     englishName:
-        "French",
+        "Portuguese",
     nativeName:
-        "Français",
+        "Português",
 
     "Language":
-        "Langue", // not used now, but will be used in options afterwards
+        "Idioma", // not used now, but will be used in options afterwards
 
     // Translation: tooltips and notification
     // (to see the notification window, remove several phrases from any dictionary, except for the English one)
     "Add more":
-        "Ajouter une traduction",
+        "Adicionar mais",
     "The translation isn't complete.":
-        "La traduction n'est pas terminée.",
+        "A tradução não está completa.",
     "The following phrases are not translated:":
-        "Les assertions suivantes n'ont pas été traduites:",
+        "As seguintes frases não são traduzidas:",
     "You can improve the translation here":
-        "Vous pouvez améliorer la traduction ici",
+        "Pode melhorar a tradução aqui",
 
     // Initial screen
     "#helpButton - learn more about the app":
-        "#helpButton - à propos",
+        "#helpButton - saber mais sobre a aplicação",
     "#optionsButton - see the available options":
-        "#helpButton - options disponibles",
+        "#optionsButton - ver as opções disponíveis",
     "powered with":
-        "basé sur",
+        "powered by",
     "Drag & Drop a file here or click to choose manually":
-        "Glisser-Déposer ici un fichier, ou cliquer pour le slectionner manuellement",
+        "Arrastar e largar um ficheiro aqui ou clicar para escolher manualmente",
     "Paste a URL to a djvu file here":
-        "Copier un lien vers un fichier .djvu",
+        "Colar aqui um URL a um ficheiro djvu",
     "Open URL":
-        "Ouvrir le lien",
+        "Abrir URL",
     'Enter a valid URL (it should start with "http(s)://")': // an alert shown when you try to open an empty URL
-        'Insérer une URL valide (doit commencer par "http(s)://")',
+        'Introduza um URL válido (deve começar com "http(s)://")',
 
     // Errors. Usually there is a header and a message for each error type.
     // For the web request error there are different types of messages depending on the HTTP status.
     // The ways to see the errors in the viewer are described in comments below.
     // In case of web requests you can load links via the browser extension (via the URL field on the initial screen)
     "Error":
-        "Erreur",
+        "Erro",
     "Error on page":
-        "Erreur dans la page", // Open 'library/assets/czech_indirect/index.djvu
+        "Erro na página", // Open 'library/assets/czech_indirect/index.djvu
     "Network error":
-        "Erreur de réseau", // Disable internet connection and try to load something by URL
+        "Erro de rede", // Disable internet connection and try to load something by URL
     "Check your network connection":
-        "Vérifier votre connexion internet",
+        "Verifique a sua ligação à rede",
     // Load any URL to a nonexistent page on the Internet,
     // e.g. https://djvu.js.org/nonexistentpage
     "Web request error":
-        "Erreur de requête web",
+        "Pedido de erro na Web",
     "404 Document not found":
-        "404 Dcument non trouvé",
+        "404 Documento não encontrado",
     "403 Access forbidden":
-        "403 Accès interdit",
+        "403 Acesso proibido",
     "500 Internal server error":
-        "500 Erreur interne du serveur",
+        "500 Erro interno do servidor",
     "The request failed with HTTP status #status":
-        "La requête a échouée avec le statut HTTP #status",
+        "O pedido falhou com o status HTTP #status",
     "DjVu file is corrupted": // Open "/library/assets/czech_indirect/dict0085.iff"
-        "Le fichier DjVu est corrompu",
+        "O ficheiro DjVu está corrompido",
     "The file doesn't comply with the DjVu format specification or it's not a whole DjVu document":
-        "Le fichier ne respecte pas la spécification du format DjVu ou il s'agit d'un document DjVu incomplet",
+        "O ficheiro não cumpre as especificações do formato DjVu ou não é um documento DjVu completo",
     "Incorrect file format": // Open a not-djvu file.
-        "Format de fichier incorrect",
+        "Formato de ficheiro incorrecto",
     "The provided file is not a DjVu document":
-        "Le fichier fourni n'est pas un document DjVu",
+        "O ficheiro fornecido não é um documento DjVu",
     // Load a URL to a DjVu file with "#page=100500" at the end (both in continuous scroll and single-page view modes)
     // e.g. https://djvu.js.org/assets/djvu_examples/DjVu3Spec.djvu#page=100500
     "Incorrect page number":
-        "Numéro de page incorrect",
+        "Número de página incorrecto",
     "There is no page with the number #pageNumber":
-        "Il n'existe pas de page avec le numéro #pageNumber",
+        "Não há página com o número #pageNumber",
     // "baseURL" is a URL to a document directory,
     // all links inside the document index.djvu are considered relative to this URL.
     // The term "base URL" can be translated as "a URL to the document's folder".
     "No base URL for an indirect DjVu document":  // Open "/library/assets/czech_indirect/index.djvu"
-        "URL de base manquante du document DjVu indirect (multi-fichier)",
+        "Sem URL base para um documento indirecto DjVu",
     "You probably opened an indirect (multi-file) DjVu document manually.":
-        "Vous avez probablement ouvert manuellement un document DjVu indirect (multi-fichier).",
+        "Provavelmente abriu manualmente um documento DjVu indirecto (multi-arquivos).",
     "But such multi-file documents can be only loaded by URL.":
-        "Mais de tels documents multi-fichier ne peuvent être ouvert que par URL",
+        "Mas tais documentos multi-arquivos só podem ser carregados por URL.",
     "Unexpected error": // Of course there is no standard way to produce this kind of error
-        "Erreur inconnue",
+        "Erro inesperado",
     "Cannot print the error, look in the console":
-        "Impossible de rapporter l'erreur, regarder dans la console.",
+        "Não é possível imprimir o erro, procurar na consola",
 
     // Options and its tooltips
     "Options":
-        "Options",
+        "Opções",
     "Show options window":
-        "Afficher la fenêtre d'option",
+        "Mostrar janela de opções",
     "Color theme":
-        "Changer de thème",
+        "Tema de cor",
     "Extension options":
-        "Options de l'extension", // the options of the browser extension
+        "Opções de extensão", // the options of the browser extension
     "Open all links with .djvu at the end via the viewer":
-        "Ouvrir tous les liеns .djvu avec l'extension",
+        "Abrir todas as ligações com .djvu no final através do visualizador",
     "All links to .djvu files will be opened by the viewer via a simple click on a link":
-        "Tous les liens .djvu seront ouverts avec l'extension en cliquant simplement sur le lien",
+        "Todos os links para ficheiros .djvu serão abertos pelo espectador através de um simples clique num link",
     "Detect .djvu files by means of http headers":
-        "Identifier les fichiers .djvu par leur en-tête http",
+        "Detectar ficheiros .djvu por meio de cabeçalhos http",
     "Analyze headers of every new tab in order to process even links which do not end with the .djvu extension":
-        "Analyser les en-têtes de chaque nouvel onglet pour identifier les fichiers même sans l'extension .djvu en fin de lien",
+        "Analisar os cabeçalhos de cada novo separador a fim de processar até ligações que não terminam com a extensão .djvu",
 
     // Footer: status bar
     "Ready":
-        "Prêt",
+        "Preparado",
     "Loading":
-        "Chargement",
+        "Carregando",
 
     // Footer: buttons' tooltips
     "Show help window":
-        "Afficher la fenêtre d'aide",
+        "Mostrar janela de ajuda",
     "Switch full page mode":
-        "Passer en mode pleine page",
+        "Mudar o modo de página inteira",
 
     // File Block tooltips
     "Choose a file":
-        "Choisir un fichier",
+        "Escolha um ficheiro",
     "Close document":
-        "Fermer le document",
+        "Fechar documento",
     "Save document":
-        "Enregistrer le document",
+        "Guardar documentação",
     "Save":
-        "Enregistrer",
+        "Guardar",
     "Open another .djvu file":
-        "Ouvrir un autre fichier .djvu",
+        "Abrir outro ficheiro .djvu",
 
     // Help window
     "The application for viewing .djvu files in the browser.":
-        "Une application pour afficher des fichiers .djvu dans le naviguateur.",
+        "A aplicação para visualizar ficheiros .djvu no browser.",
     "If something doesn't work properly, feel free to write about the problem at #email.":
-        "Si quelque chose ne fonctionne pas correctement, vous pouvez écrire à #email.",
+        "Se algo não funcionar correctamente, sinta-se à vontade para escrever sobre o problema em #email.",
     "The official website is #website.":
-        "Le site officiel est #website.",
+        "O site oficial é #website.",
     "The source code is available on #link.":
-        "Le code source est disponible à l'adresse: #link.",
+        "O código fonte está disponível em #link.",
     "Hotkeys":
-        "Raccourcis Clavier",
+        "Teclas de atalho",
     "save the document":
-        "enregistrer le document",
+        "guardar o documento",
     "go to the previous page":
-        "page précédente",
+        "ir para a página anterior",
     "go to the next page":
-        "page suivante",
+        "ir para a página seguinte",
     "Controls":
-        "Boutons",
+        "Controles",
     "#expandIcon and #collapseIcon are to switch the viewer to the full page mode and back.":
-        "#expandIcon & #collapseIcon permettent de passer en mode pleine page et inversement.",
+        "#expandIcon e #collapseIcon devem mudar o visualizador para o modo página inteira e voltar.",
     "If you work with the browser extension, these buttons will cause no effect, since the viewer takes the whole page by default.":
-        "Si vous utilisez l'extension pour navigateur, ces boutons n'auront aucun effet, car l'application occupe déjà toute la page.",
+        "Se trabalhar com a extensão do navegador, estes botões não causarão qualquer efeito, uma vez que o visualizador leva a página inteira por defeito.",
 
     // Toolbar tooltips
     "Continuous scroll view mode":
-        "Mode défilement continu",
+        "Modo de visualização scroll contínuo",
     "Single page view mode":
-        "Mode page unique",
+        "Modo de visualização de uma página",
     "Text view mode":
-        "Mode texte",
+        "Modo de visualização de texto",
     "Click on the number to enter it manually":
-        "Cliquer sur le nombre pour en saisir un manuellement",
+        "Clique no número para o introduzir manualmente",
     "Rotate the page":
-        "Faire pivoter la page",
+        "Rodar a página",
     "You also can scale the page via Ctrl+MouseWheel":
-        "Vous pouvez également ajuster la page avec Ctrl+Molette",
+        "Também pode escalar a página através de Ctrl+MouseWheel",
     "Text cursor mode":
-        "Curseur pour mettre le texte en surbrillance",
+        "Modo cursor de texto",
     "Grab cursor mode":
-        "Défilement glisser-déposer",
+        "Modo de agarrar o cursor",
 
     // Contents
     "Contents":
-        "Table des matières",
+        "Conteúdos",
     "No contents provided":
-        "Pas de sommaire",
+        "Nenhum conteúdo fornecido",
     // A rare case. Open /library/assets/links.djvu in the viewer on https://djvu.js.org/ (not in the extension!)
     // and click the "Absolute Link" in the contents
     "The link points to another document. Do you want to proceed?":
-        "Le lien redirige vers un autre document. Voulez-vous continuer ?",
+        "A ligação aponta para outro documento. Quer prosseguir?",
 
     // Text Block (shown in the text view mode)
     "No text on this page":
-        "Pas de texte dans cette page",
+        "Nenhum texto nesta página",
 
     // Save dialog (shows when you save an indirect djvu)
     "You are trying to save an indirect (multi-file) document.":
-        "Vous essayez d'enregistrer un document multi-fichier ",
+        "Está a tentar salvar um documento indirecto (multi-arquivo).",
     "What exactly do you want to do?":
-        "Que voulez-vous faire ?",
+        "Que queres fazer exactamente?",
     "Save only index file":
-        "Enregistrer uniquement le fichier racine",
+        "Guardar só ficheiro de índice",
     "Download, bundle and save the whole document as one file":
-        "Télécharger, rassembler & sauveguarder l'ensemble du document dans un seul fichier",
+        "Descarregar, agrupar e guardar o documento inteiro como um só ficheiro",
     "Downloading and bundling the document":
-        "Téléchargement & assemblage du document",
+        "Descarregar e empacotar o documento",
     "The document has been downloaded and bundled into one file successfully":
-        "Le document a été téléchargé, rassemblé & sauveguardé dans un seul fichier avec succès",
+        "O documento foi descarregado e agrupado num único ficheiro com sucesso",
 
     // Printing
     "Print document":
-        "Imprimer le document",
+        "Imprimir documento",
     "Pages must be rendered before printing.":
-        "Les pages doivent être rendues avant impression",
+        "As páginas devem ser entregues antes da impressão.",
     "It may take a while.":
-        "Cela peut prendre un moment",
+        "Pode demorar algum tempo.",
     "Select the pages you want to print.":
-        "Sélectionner la page à imprimer",
+        "Seleccione as páginas que pretende imprimir.",
     "From":
         "De",
     "to":
-        "à",
+        "a",
     "Prepare pages for printing":
-        "Préparer les pages pour l'impression",
+        "Preparar páginas para impressão",
     "Preparing pages for printing":
-        "Préparation des pages pour l'impression",
+        "Preparar páginas para impressão",
 };