+++ viewer/src/locales/index.js
@@ -8,6 +8,8 @@ import Swedish from './Swedish.js';
 import French from "./French.js";
 import Italian from "./Italian.js";
 import ChineseSimplified from "./ChineseSimplified.js"
+import Spanish from "./Spanish.js";
+import Portuguese from "./Portuguese.js";
 
 /**
  * Here we use 2-character lowercase ISO 639-1 codes.
@@ -21,4 +23,6 @@ export default {
     'fr': French,
     'it': Italian,
     'zh': ChineseSimplified,
+    'pt': Portuguese,
+    'es': Spanish,
 };